refactor(app): tighten external seam exports
